itachi.memory-arc/requirements.txt
# Core dependencies (minimal for basic functionality)
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Optional: Enhanced heuristics (not required for basic usage)
# Uncomment if you want enhanced heuristic features
# keybert>=0.8.0
# spacy>=3.7.0

# Optional: AI adapters (install only what you need)
# Uncomment the adapters you want to use
# openai>=1.0.0
# anthropic>=0.18.0
# ollama>=0.1.0
# huggingface-hub>=0.20.0
